Separate Type into it's own file
